app/requirements.txt
Flask==2.3.2
joblib==1.4.0
pandas==2.2.2
scikit-learn==1.5.0
PyYAML==6.0.1
